ref: Extract transport layer into dedicated module
Move transport-related types (BoxTransport, ManagedTransport,
ChannelManager, ChannelStream) from bigtable.rs into a new
transport.rs module. Introduce ManagedTransportBuilder for
configuring managed channel pools, and a generic new_with_transport
constructor that all existing codepaths now use internally.

Breaking changes:
- new_with_token_provider and new_with_emulator are now async
- new_with_managed_transport is removed (use ManagedTransportBuilder
  + new_with_transport instead)
- BoxTransport moved from crate::bigtable to crate::transport
